requirements.txt
astroid==2.4.2
Django==2.2
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
flake8==3.8.3
flake8-django==1.1.1
importlib-metadata==2.0.0
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.2.0
